# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Melogold API",
    "version": "1",
    "description": "Sync server of Melogold: favorites, library, playlists, history and playback across a user's devices. The normative contract is docs/API.md; clients branch only on `code` of ErrorResponse (error-codes.json). There is no version in the URL: compatibility comes from `apiVersion`/`minApiVersion` and `features.sync.protocol` of `GET /server/info` (API §1.1).",
    "license": {
      "name": "CC0-1.0",
      "url": "https://creativecommons.org/publicdomain/zero/1.0/"
    }
  },
  "servers": [
    {
      "url": "https://api.melogold.app",
      "description": "The official server. Self-hosted servers serve the same API under their own base URL."
    }
  ],
...
  "paths": {
...
    "/auth/link/requests": {
      "post": {
        "operationId": "createLinkRequest",
        "summary": "New device: start a link and show a QR code (mode request)",
        "tags": [
          "linking"
        ],
        "description": "At most 20 active requests per IP (API §1.10). Not retried automatically.\n\nRate limit: 30/10 min per IP (IPv6 by /56) (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkRequestRequest"
              }
            }
          }
        },
        "security": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCreated"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/link/claim": {
      "post": {
        "operationId": "claimLink",
        "summary": "New device: claim an invite scanned from a signed-in device (mode invite)",
        "tags": [
          "linking"
        ],
        "description": "Show `verifyCode` large; the signed-in device approves with it (API §4.6). Not retried.\n\nRate limit: 30/10 min per IP (IPv6 by /56) (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimLinkRequest"
              }
            }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkClaimed"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`link_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`link_already_claimed`, `link_wrong_mode`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "`link_expired`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/link/poll": {
      "post": {
        "operationId": "pollLink",
        "summary": "New device: long-poll the link until the session is issued",
        "tags": [
          "linking"
        ],
        "description": "Answers at once when the status differs from `knownStatus`, otherwise waits up to `waitSeconds` (25). After `completed` the same poll returns the same session for 60 s. HTTP client timeout: 35 s (API §4.6).\n\nRate limit: 60/min per poll secret and 600/min per IP (IPv6 by /56) (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollLinkRequest"
              }
            }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkPollResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`link_denied`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`link_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`device_limit_reached` (deviceLimit, deviceCount)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "`link_expired`, `link_cancelled`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/link/cancel": {
      "post": {
        "operationId": "cancelLinkRequest",
        "summary": "New device: cancel the link",
        "tags": [
          "linking"
        ],
        "description": "Rate limit: 60/min per poll secret and 600/min per IP (IPv6 by /56) (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelLinkRequestRequest"
              }
            }
          }
        },
        "security": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`link_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/me/links": {
      "post": {
        "operationId": "createLinkInvite",
        "summary": "Signed-in device: create an invite and show its QR code (mode invite)",
        "tags": [
          "linking"
        ],
        "description": "`pollSecret` of the answer is null. A fourth active invite cancels the oldest (API §4.6).\n\nRate limit: 10/10 min per user (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkInviteRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCreated"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me/links/resolve": {
      "post": {
        "operationId": "resolveLink",
        "summary": "Signed-in device: open a request by its QR token or user code (mode request)",
        "tags": [
          "linking"
        ],
        "description": "Exactly one of `linkToken` and `userCode`. The answer has status claimed and verifyChoices.\n\nRate limit: 20/10 min per user (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveLinkRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkDetails"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`link_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`link_already_claimed`, `link_wrong_mode`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "`link_expired`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me/links/{linkId}": {
      "get": {
        "operationId": "getLink",
        "summary": "Signed-in device: the approval card of a link",
        "tags": [
          "linking"
        ],
        "description": "Rate limit: 120/min per user (API §1.10).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "linkId",
            "required": true,
            "description": "UUID, lowercase only (API §1.6)."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkDetails"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`link_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/me/links/{linkId}/cancel": {
      "post": {
        "operationId": "cancelLink",
        "summary": "Signed-in device: cancel its invite",
        "tags": [
          "linking"
        ],
        "description": "Rate limit: 120/min per user (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "linkId",
            "required": true,
            "description": "UUID, lowercase only (API §1.6)."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`link_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token of TokenPair (API §1.7)."
      }
    },
    "schemas": {
...
      "CancelLinkRequestRequest": {
        "type": "object",
        "properties": {
          "pollSecret": {
            "type": "string",
            "pattern": "^mgps_[A-Za-z0-9_-]{43}$",
            "description": "`mgps_` + 43 characters (API §1.6)."
          }
        },
        "required": [
          "pollSecret"
        ]
      },
...
      "ClaimLinkRequest": {
        "type": "object",
        "properties": {
          "linkToken": {
            "nullable": true,
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$",
            "description": "43 characters of base64url (API §1.6)."
          },
          "userCode": {
            "nullable": true,
            "description": "Crockford Base32, shown as XXXX-XXXX. Input: case-insensitive; spaces, `-`, `_` ignored; O→0, I and L→1; exactly 8 characters.",
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceInput"
          }
        },
        "required": [
          "device"
        ],
        "description": "Exactly one of linkToken and userCode."
      },
...
      "CreateLinkInviteRequest": {
        "type": "object",
        "properties": {},
        "description": "An empty JSON object."
      },
      "CreateLinkRequestRequest": {
        "type": "object",
        "properties": {
          "device": {
            "$ref": "#/components/schemas/DeviceInput"
          }
        },
        "required": [
          "device"
        ]
      },
...
      "EmptyRequest": {
        "type": "object",
        "properties": {},
        "description": "An empty JSON object."
      },
...
    }
  }
}